chore: 백엔드 설정 및 데이터 구조 개선
- application.properties 설정 최적화
- SpecialDayCategoryRepository 메서드 추가
- CalendarService 및 UserController 로직 보완
- CalendarDto 데이터 필드 확장

+++ backend/src/main/java/com/voiz/mapper/SpecialDayCategoryRepository.java
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.voiz.vo.SpecialDayCategory;
 
 @Repository
-public interface SpecialDayCategoryRepository extends JpaRepository<SpecialDayCategory, Long> {
+public interface SpecialDayCategoryRepository extends JpaRepository<SpecialDayCategory, Integer> {
     
     /**
      * 특정 특일의 모든 카테고리 조회
